sort_done

+++ ds/source/sorts/sort.h
@@ -1,9 +1,9 @@
 #ifndef __ILRD_CR02_SORT__
 #define __ILRD_CR02_SORT__
 
-void BubbleSort(int arr[], int len);
+void BubbleSort(int arr[], size_t len);
 
-void SelectionSort(int arr[], int len);
+void SelectionSort(int arr[], size_t len);
 
 void InsertionSort(int arr[], int len);
